+++ src/se/devscout/android/model/repo/sql/cache/IdCacheEntry.java
@@ -1,11 +1,11 @@
 package se.devscout.android.model.repo.sql.cache;
 
 public class IdCacheEntry {
-    private final long[] mCompareValues;
+    private final double[] mCompareValues;
     private long mId;
     private long mServerId;
 
-    public IdCacheEntry(long id, long serverId, long[] values) {
+    public IdCacheEntry(long id, long serverId, double[] values) {
         mId = id;
         mServerId = serverId;
         mCompareValues = values;
@@ -21,7 +21,7 @@ public class IdCacheEntry {
 
     public boolean isAdditionalValuesListIdentical(IdCacheEntry idCacheEntry) {
         for (int i = 0; i < mCompareValues.length; i++) {
-            long value = mCompareValues[i];
+            double value = mCompareValues[i];
             if (value != idCacheEntry.mCompareValues[i]) {
                 return false;
             }
